Proposing "mattermost" -> "Mattermost" tweak?
@lazyfrosch, thanks so much for starting this project! Proposing just a small PR so Mattermost is capitalized consistently (sort of like MySQL, we more commonly write "Mattermost" when it's not a variable)

This is wonderful work, we're so glad to have you involved. 

RC1 for our next release ships on Monday, and final release is Oct 2. 

If you're interested in that release there will be a new config file to swap out, along with the new bits, and that release brings a host of improvements, like an admin console, [markdown support](http://www.mattermost.org/open-source-slack-alternative-adopts-markdown/) and new theme colors. 

We hope you like it!